* libctl/ml_bidi.o, libctl/ml_iscii.o: Memory size passed to malloc() is fixed   in ml_bidi_new() and ml_iscii_new(). * mlterm/Makefile.in: ml_bidi.o and ml_iscii.o are moved from OBJ_CORE to OBJ. * x_window.h: xft_draw and cairo_draw are never removed from x_window_t. * im_scim.cpp: KeyEvent::layout is initialized explicitly in im_scim_key_event().

+++ mlterm/ml_model.h
@@ -18,8 +18,8 @@ typedef struct  ml_model
 	ml_line_t *  lines ;
 	
 	/* public(readonly) */
-	u_int  num_of_cols ;
-	u_int  num_of_rows ;
+	u_int16_t  num_of_cols ;	/* 0 - 65536 */
+	u_int16_t  num_of_rows ;	/* 0 - 65536 */
 	
 	/* private */
 	int  beg_row ;			/* used for scrolling */
